% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMEND.R
\name{core_cc}
\alias{core_cc}
\title{Calculate core-clustering coefficients}
\usage{
core_cc(g, weight = TRUE)
}
\arguments{
\item{g}{igraph. Must be connected}

\item{weight}{Logical. If true, the edge density is calculated taking into account edge weights}
}
\value{
Vector of core-clustering coefficients for each node
}
\description{
\code{core_cc()} calculates the core-clustering coefficients of all nodes in a graph. This network concept was introduced by \emph{Bader & Hogue} (2003).
}
\details{
The core-clustering coefficient of a node is the edge density of the maximum k-core of the immediate neighborhood of that node. The k-core of a graph is the maximal subgraph in which every node has degree >= k.
}
\examples{
library(igraph)

graph <- sample_pa(n = 100, power = 1.2)
core_cc(graph)

}
