% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integrated_network.R
\name{extract_string}
\alias{extract_string}
\title{Extract a substring}
\usage{
extract_string(x, k, pos)
}
\arguments{
\item{x}{character vector}

\item{k}{delimiter to split on}

\item{pos}{position of split to extract. 0 to extract all splits}
}
\value{
character vector
}
\description{
Extract a substring from a specified position (\emph{pos}) after splitting on a given delimiter (\emph{k}).
}
\examples{
x <- paste(letters, LETTERS, sep = "|")
extract_string(x, "\\\\|", 2)
extract_string(x, "\\\\|", 0)

}
