% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMEND.R
\name{get_subnetwork}
\alias{get_subnetwork}
\title{Get a larger subnetwork from previous iterations of the AMEND algorithm}
\usage{
get_subnetwork(ig, amend_object, k)
}
\arguments{
\item{ig}{Input network}

\item{amend_object}{Result from \code{module_identification()}}

\item{k}{Iteration associated with the subnetwork you want to retrieve}
}
\value{
igraph object of the subnetwork associated with the given iteration.
}
\description{
This function allows the user to examine intermediate subnetworks that were generated during \code{module_identification()}, since the algorithm may overshoot the approximate final module size, or the user may be interested in the parent subnetworks of the final module.
}
\details{
Since \strong{AMEND} iteratively filters out nodes to arrive at a final module, there are intermediate subnetworks associated with each iteration. A subnetwork at an earlier iteration contains the subnetworks of all later iterations. This function can be used to investigate these earlier subnetworks, for example, to identify when in the algorithm certain nodes were removed.
}
\examples{
# Attach igraph library
library(igraph)

# Perform module identification 
mod <- module_identification(network_layers = simple_network, n = 25)

# Get the processed igraph of the input network
g <- create_integrated_network(network_layers = simple_network)

# Get the subnetwork from the end of the 4th iteration of the AMEND algorithm
get_subnetwork(ig = g$network, amend_object = mod, k = 4)

}
