% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interlayer_links}
\alias{interlayer_links}
\title{Interlayer links for multilayer network}
\format{
list with names 'protdat|phosphdat', 'prot|meta', 'prot2|meta2', 'metadat|metadat2' and 'protdat|protdat2'.
}
\source{
\url{http://stitch.embl.de/}
}
\usage{
interlayer_links
}
\description{
\code{interlayer_links} is a list of 3 character scalars of keyword "common" and 2 igraph objects containing interlayer links. Chemcial-protein interactions (i.e., interlayer links) were obtained from STITCH (v5.0).
}
\keyword{datasets}
