% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalk.R
\name{ipf}
\alias{ipf}
\title{Iterative Proportional Fitting}
\usage{
ipf(X, gamma = 1, diag_weight = 1e-08)
}
\arguments{
\item{X}{Matrix}

\item{gamma}{factor controlling to what extent the transition matrix will be adjusted to mitigate degree influence on diffusion processes. Between 0 (no adjustment) and 1 (maximum adjustment).}

\item{diag_weight}{Edge weight to add to zero-value diagonal elements of X to aid on convergence}
}
\value{
A named list
B: scaled matrix from the matrix product PXQ
p: diagonal elements of diagonal matrix P
q: diagonal elements of diagonal matrix Q

@noRd
}
\description{
Scales input matrix X to have row and column sums approximately equal to 1, thereby transforming X to a bistochastic (doubly stochastic) matrix.
}
