% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multilayer_data_runif}
\alias{multilayer_data_runif}
\title{Node-wise data for multilayer network}
\format{
list of numeric vectors with names 'metadat', 'protdat', 'phosphdat', 'metadat2' and 'protdat2'.
}
\usage{
multilayer_data_runif
}
\description{
\code{multilayer_data_runif} is a list containing vectors of data randomly generated from a Uniform(0,1) distribution representing p-values from some 'omic experiment.
}
\keyword{datasets}
