% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multilayer_hierarchy}
\alias{multilayer_hierarchy}
\title{Hierarchy for multilayer network}
\format{
2-column matrix of a directed edgelist
}
\usage{
multilayer_hierarchy
}
\description{
\code{multilayer_hierarchy} is a 2-column matrix representing a directed edgelist for the hierarchy of the multilayer network given in \code{multilayer_network}.
}
\keyword{datasets}
