% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multilayer_network}
\alias{multilayer_network}
\title{Multilayer Network}
\format{
list of 5 igraph objects with names 'metadat', 'protdat', 'phosphdat', 'metadat2' and 'protdat2'.
}
\source{
\url{https://string-db.org/}, \url{http://stitch.embl.de/}
}
\usage{
multilayer_network
}
\description{
\code{multilayer_network} is a list of 5 igraph objects representing layers of a multilayer network. Protein-protein interactions (PPIs) were obtained from STRING (v12.0), and chemical-chemical & chemcial-protein interactions were obtained from STITCH (v5.0).

The 5 igraph objects were obtained by getting clusters (Louvain clustering method) for STRING and STITCH networks and taking the two largest clusters from each network. One cluster is included twice to represent another layer.
}
\keyword{datasets}
