% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AMEND.R
\name{node_connectivity_score}
\alias{node_connectivity_score}
\title{Calculate a node-wise connectivity score}
\usage{
node_connectivity_score(
  graph,
  net_diam_prop = -1,
  inverse = FALSE,
  mode = c("core", "mix")
)
}
\arguments{
\item{graph}{igraph}

\item{net_diam_prop}{numeric. Proportion of graph to use when calculating betweenness and harmonic centrality, for mode="mix" only.}

\item{inverse}{Logical. Whether to return values inversely related to connectivity measure selected in \strong{mode}.}

\item{mode}{character scalar. One of "core" or "mix". Default is "core"}
}
\value{
numeric vector
}
\description{
Calculates one or several centrality metrics and converts to empirical cumulative probabilities. Used to set node-specific restart values in RWR (with inverse=TRUE).
}
\details{
For mode="core", only considers the coreness of nodes when calculating ECDF. For mode="mix", it considers a mix of betweenness centrality, harmonic centrality, and core clustering coefficient, aggregated by applying ECDF to each score individually then taking a weighted average of these probabilities.
When inverse=TRUE, multiplies values by -1 before ECDF calculation.
}
\examples{
net <- create_integrated_network(network_layers = simple_network, lcc = TRUE)$network
ncs <- node_connectivity_score(graph = net, inverse = FALSE, mode = "core")
ncs_inv <- node_connectivity_score(graph = net, inverse = TRUE, mode = "core")
node_coreness <- core_cc(g = net)
cor(node_coreness, ncs)
cor(node_coreness, ncs_inv)

}
