% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomWalk.R
\name{seed_weight_sets}
\alias{seed_weight_sets}
\title{Possible seed_weight parameters}
\usage{
seed_weight_sets(hierarchy)
}
\arguments{
\item{hierarchy}{igraph of network hierarchy}
}
\value{
list. list element will be an empty list if no seed weight parameters are required.
}
\description{
Get sets of siblings in hierarchy that can have seed weight parameters. These will be any sibling sets with two or more nodes.
}
\details{
This function may be useful to determine which sets of categories in your network hierarchy should be given seed weights for use in \code{RWR()}.
}
\examples{
# Attach igraph package
library(igraph)

# Inspect 'multilayer_hierarchy' provided in HMNM package
multilayer_hierarchy

# Create network hierarchy as an igraph object with additional class 'hierarchy'
net_hier <- create_network_hierarchy(multilayer_hierarchy)

# Get the sibling sets of hierarchy
seed_weight_sets(net_hier)

}
\seealso{
\code{\link[=RWR]{RWR()}}, \code{\link[=RWR_pipeline]{RWR_pipeline()}}, \code{\link[=module_identification]{module_identification()}}
}
