% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simple_network}
\alias{simple_network}
\title{Simple Network}
\format{
3-column matrix represnting an edgelist with an additional columm of edge weights.
}
\source{
\url{https://string-db.org/}
}
\usage{
simple_network
}
\description{
\code{simple_network} is a 3-column matrix represnting an edgelist with an additional columm of edge weights. These edges are protein-protein interactions (PPIs) obtained from STRING (v12.0).
}
\keyword{datasets}
